DO NOT MERGE: this commit should fail the tests (but doesn't)
Changing the error produced by the EidasValidationPartialController should cause 'When the user visits the redirect to country page' to fail, since it now goes to a different error page with different content. However, the test continues to pass. This shows that we are getting the 'something has gone wrong' page, but for the wrong reasons...